app/src/test/java/br/com/alura/leilao/api/EnviadorDeLanceTest.java
package br.com.alura.leilao.api;

import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.doThrow;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.verify;

import android.content.Context;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mock;
import org.mockito.junit.MockitoJUnitRunner;

import br.com.alura.leilao.api.retrofit.client.LeilaoWebClient;
import br.com.alura.leilao.exception.UsuarioJaDeuCincoLancesException;
import br.com.alura.leilao.model.Lance;
import br.com.alura.leilao.model.Leilao;
import br.com.alura.leilao.model.Usuario;
import br.com.alura.leilao.ui.dialog.AvisoDialogManager;

@RunWith(MockitoJUnitRunner.class)
public class EnviadorDeLanceTest {

    @Mock
    private LeilaoWebClient client;
    @Mock
    private EnviadorDeLance.LanceProcessadoListener listener;
    @Mock
    private AvisoDialogManager aviso;

    @Test
    public void deve_MostrarMensagemDeFalha_QuandoLanceForMenorQueUltimoLance() {
        EnviadorDeLance enviadorDeLance = new EnviadorDeLance(client, listener, aviso);
        Leilao leilao = new Leilao("Console");
        Lance joao = new Lance(new Usuario("Joao"), 200.0);

        leilao.propoe(joao);
        enviadorDeLance.envia(leilao, new Lance(new Usuario("Joana"), 100.0));

        verify(aviso).mostraAvisoLanceMenorQueUltimoLance();
    }

    @Test
    public void deve_MostrarMensagemDeFalha_QuandoUsuarioDerMaisDeCincoLances() {
        EnviadorDeLance enviadorDeLance = new EnviadorDeLance(client, listener, aviso);
        Leilao leilao = mock(Leilao.class);
        doThrow(UsuarioJaDeuCincoLancesException.class)
                .when(leilao).propoe(any(Lance.class));

        enviadorDeLance.envia(leilao, new Lance(new Usuario("Joao"), 200.0));

        verify(aviso).mostraAvisoUsuarioJaDeuCincoLances();
    }

}